Time: 791 ms (5.04%), Space: 80 MB (6.64%) - LeetHub

+++ 0204-count-primes/0204-count-primes.java
@@ -12,8 +12,8 @@ class Solution {
         //         cnt++;
         // }
         // if (n>1)
-        for (long i = 2; i * i <= n; i++) {
-            if (prime[(int) i]) {
+        for (int i = 2; i * i <= n; i++) {
+            if (prime[i]) {
                 for (long j = (long) i * i; j < n; j += i) {
                     prime[(int) j] = false;
